moves MockShardedQueryable test helper to queryrange pkg
